Lap and teleport system
Turns out the portal events don't trigger if your on a horse so I used a move event that just checks your distance to the portal. Then there was some weird issue with ejecting the player from the horse and teleporting so I am now killing the horse when you enter the portal, and then spawning a new one when you teleport back to the starting position. Works well.